support observer abort / reevaluating whenever any utilities change
The Utility composite now tells UtilityBlackboard decorators running on children to request execution (retrigger a search) when the blackboard value (utility) they're watching changes.
This causes the Utility composite node to be hit which recalculates utility.
The Utility composite has to set this up because its the one that's always active. However all its functions are const so it can't pass itself to the blackboard key notification setup function, so it uses the child decorator nodes which aren't const.